This commit contains some major changes to sims_ocs and Scheduler (#115)
This commit contains some major changes to sims_ocs to update the way it communicates with the scheduler. Previous behavior was that SOCS would read the configuration and send the parameters through DDS to the Scheduler. But this is not the way it is going to work on real operations. The new behavior implemented here is that SOCS will get a “tag” with the configuration name and send this over DDS to the Scheduler. The Scheduler translates that tag into a set of configurations, read them and send them back over DDS.

- make use of the commandable DDS logic to synchronize SOCS and Scheduler. This way, each party can wait for the other to be ready and take the appropriate action.

- use new set of DDS topics where target and other high priority information is passed as events rather than telemetry.